+++ proofs/benches/zswap_output.rs
@@ -268,8 +268,8 @@ fn bench_zswap_output(c: &mut Criterion) {
         std::slice::from_ref(&circuit),
         1,
         &[&[&[], &instance]],
-        OsRng,
         &mut transcript,
+        &mut OsRng,
         &mut group,
     )
     .expect("Failed to generate proof");