+++ proofs/examples/serialization.rs
@@ -161,8 +161,8 @@ fn main() {
         #[cfg(feature = "committed-instances")]
         0,
         &[instances],
-        OsRng,
         &mut transcript,
+        OsRng,
     )
     .expect("proof generation should not fail");