+++ proofs/src/poly/kzg/params.rs
@@ -127,7 +127,7 @@ impl<E: Engine + Debug> ParamsKZG<E> {
         self.g2
     }
 
-    /// Returns first power of secret on G2
+    /// Returns \[τ\]₂, a commitment to τ in G2.
     pub fn s_g2(&self) -> E::G2 {
         self.s_g2
     }
@@ -242,6 +242,11 @@ impl<E: MultiMillerLoop + Debug> ParamsVerifierKZG<E>
 where
     E::G2: Curve + ProcessedSerdeObject,
 {
+    /// Returns \[τ\]₂, a commitment to τ in G2.
+    pub fn s_g2(&self) -> E::G2 {
+        self.s_g2
+    }
+
     /// Writes parameters to buffer
     pub fn write<W: io::Write>(&self, writer: &mut W, format: SerdeFormat) -> io::Result<()> {
         self.s_g2.write(writer, format)?;